Sales Assistant/requirements.txt
streamlit
streamlit-webrtc
deepgram-sdk
websockets
requests
python-dotenv
